requirements.txt
pytest
requests
allure-pytest==2.15.0